Se implementaron las funciones para el CRUD de producto, además de la Exportacion a PDF

+++ src/main/java/com/example/demo/repository/ProductoRepository.java
@@ -6,7 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.ProductoEntity;
 
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Long>{
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer>{
+
+
